requirements.txt
# ML
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
xgboost>=2.0
lightgbm>=4.0
matplotlib>=3.7
seaborn>=0.12
joblib>=1.3

# NLP
spacy>=3.7
pdfplumber>=0.10
pytesseract>=0.3
Pillow>=10.0

# API
fastapi>=0.104
uvicorn>=0.24
python-multipart>=0.0.6
python-dotenv>=1.0

# Dev
pytest>=7.4
black>=23.0
